Slowly progress in the handling of root_intr, based on ohci example
XXX: I'm not sure that dwc_otg_root_intr must be called directly. In
ohci / ehci, the equivalent function is called in softint context, while
in uhci, it is called in an callout context.
Moreover, for moment, the method expects that the lock are correctly
helded.

dwc_otg_intr1 must be completed by calling it (or some wrappers)

while here, fix a typo in a config description.